t5 force to graph mode

+++ examples/opensora_pku/opensora/sample/sample_t2v.py
@@ -348,9 +348,11 @@ if __name__ == "__main__":
                 loaded_latents.append(np.load(save_fp))
             loaded_latents = np.stack(loaded_latents)
             decode_data = vae.decode(ms.Tensor(loaded_latents))
-            decode_data = ms.ops.clip_by_value(
-                (decode_data + 1.0) / 2.0, clip_value_min=0.0, clip_value_max=1.0
-            ).asnumpy()
+            decode_data = (
+                ms.ops.clip_by_value((decode_data + 1.0) / 2.0, clip_value_min=0.0, clip_value_max=1.0)
+                .to(ms.float32)
+                .asnumpy()
+            )
             for i_sample in range(args.batch_size):
                 save_fp = os.path.join(save_dir, file_paths[i_sample]).replace(".npy", ".gif")
                 save_video_data = decode_data[i_sample : i_sample + 1].transpose(
